requirements.txt
requests
python-dotenv
faiss-cpu
langchain
langchain-community
scikit-learn
sentence-transformers
